Extract quality policy helpers to api/quality_policy.py
Move load_quality_policy, apply_quality_policy_defaults,
normalize_quality_severity, count/has quality issue helpers,
and write_quality_report_output (~200 LOC) from generator.py
to a dedicated module. Re-imported in generator.py for
backwards compatibility.
